Aula 7 - Implmentando o passport

+++ routes/tasks.js
@@ -2,9 +2,10 @@ module.exports = app =>{
     const Tasks = app.db.models.Tasks
 
     app.route('/tasks')
+        .all(app.auth.authenticate())
         .get((req, res)=>{
             // "/tasks": Lista de tarefas
-            Tasks.findAll({})
+            Tasks.findAll({where: {user_id: req.user.id}})
                 .then(result =>{
                     res.json(result)
                 })
@@ -14,6 +15,7 @@ module.exports = app =>{
         })
         .post((req, res)=>{
             // "/tasks": Cadastra uma nova tarefa
+            req.body.user_id=req.ser.id;
             Tasks.create(req.body)
             .then(result =>{
                 res.json(result)
@@ -24,9 +26,10 @@ module.exports = app =>{
         });
         
     app.route('/tasks/:id')
+        .all(app.auth.authenticate())
         .get((req, res)=>{
             // "/tasks/1": Consulta 1 tarefa
-            Tasks.findOne({where: req.params})
+            Tasks.findOne({where: {id: req.params.id, user_id: req.user.id}})
                 .then(result=>{
                     if (result){
                         res.json(result);
@@ -40,7 +43,7 @@ module.exports = app =>{
         })
         .put((req, res)=>{
             // "/tasks/1": Atualiza 1 tarefa
-            Tasks.update(req.body, {where: req.params})
+            Tasks.update(req.body, {where: {id: req.params.id, user_id: req.user.id}})
                 .then(result =>{
                     res.sendStatus(204);
                 })
@@ -50,7 +53,7 @@ module.exports = app =>{
         })
         .delete((req, res)=>{
             // "/tasks/1": Deleta 1 tarefa
-            Tasks.destroy({where: req.params})
+            Tasks.destroy({where: {id: req.params.id, user_id: req.user.id}})
                 .then(result=>{
                     res.sendStatus(204);
                 })
